Played around with EJS
Added a /video route that returns a video in .mp4 format. Added a /profile route that receives and returns profile ID route params. Installed EJS and created a Homepage that displays a location variable and a list rendered by EJS. Created a /users view that displays a user list by using EJS Includes.